requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
PyYAML
numpy==1.24.4
torchsampler
scikit-learn
#torch==1.11.0+cu113
torchvision==0.12.0
Pillow==9.5.0
fastNLP==1.0.1
transformers==4.20.1
